Add files via upload
Normal distributions were plotted as a simple demonstration for the Monte Carlo method of computational analysis. Experimental data of a cosmic muon penetrating an array of copper plates and scintillators was then simulated using the foundations of this method. This data was used to predict the outcome of the experiment. A graphical model with paths for each particle was also drawn by the code.